WindowsMessage Parameter Fix
Changed the third parameter of SendKeycodes to int so that it can take the modifiers; Also changed the ZHS translations in corresponding fields.

+++ Source/Triggernometry/Action.cs
@@ -1532,7 +1532,7 @@ namespace Triggernometry
                                         string procid = ctx.EvaluateStringExpression(ActionContextLogger, ctx, _KeyPressProcId);
                                         string window = ctx.EvaluateStringExpression(ActionContextLogger, ctx, _KeyPressWindow);
                                         int keycode = (int)ctx.EvaluateNumericExpression(ActionContextLogger, ctx, _KeyPressCode);
-                                        RealPlugin.WindowsUtils.SendKeycodes(procid, window, (ushort)keycode);
+                                        RealPlugin.WindowsUtils.SendKeycodes(procid, window, keycode);
                                     }
                                     break;
                                 case KeypressTypeEnum.WindowMessageCombo:
